ading all changes. FD Completely decoupled from common. FD configuration is successful.
