+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.12)
-project(daqconf VERSION 6.0.2)
+project(fddaqconf VERSION 6.0.2)
 
 find_package(daq-cmake REQUIRED )
