--- docs/README.md
@@ -1,7 +1,0 @@
-# fddaqconf
-
-A manual fork of the daqconf repository, intended for Far
-Detector-focused development. For daqconf documentation go
-[here](https://dune-daq-sw.readthedocs.io/en/latest/packages/daqconf/).
-
-
