--- python/daqconf/core/console.py
@@ -1,2 +1,0 @@
-from rich.console import Console
-console = Console()
